+++ demos/superjumper/src/MainMenuScreen.cpp
@@ -36,7 +36,7 @@ void MainMenuScreen::update(float deltaTime)
 {
 	if(Gdx.input->justTouched())
 	{
-		m_guiCam->unproject(m_touchPoint.set(Gdx.input->getX(), Gdx.input->getY(), 0));
+		m_guiCam->unproject(m_touchPoint.set((float)Gdx.input->getX(), (float)Gdx.input->getY(), 0));
 
 		if(OverlapTester::pointInRectangle(m_playBounds, m_touchPoint.x, m_touchPoint.y))
 		{
